+++ deployer/db/config/options.php
@@ -4,4 +4,5 @@ namespace Deployer;
 
 use Symfony\Component\Console\Input\InputOption;
 
-option('dumpcode', null, InputOption::VALUE_OPTIONAL, 'Database dump code.');
+option('dumpcode', null, InputOption::VALUE_OPTIONAL, 'Database dump code');
+option('dboptions', null, InputOption::VALUE_OPTIONAL, 'Database commands options');