+++ deployer/db/task/db_download.php
@@ -5,13 +5,14 @@ namespace Deployer;
 use SourceBroker\DeployerExtendedDatabase\Utility\ConsoleUtility;
 use SourceBroker\DeployerExtendedDatabase\Utility\FileUtility;
 use SourceBroker\DeployerExtendedDatabase\Utility\RsyncUtility;
+use Deployer\Exception\GracefulShutdownException;
 
 /*
  * @see https://github.com/sourcebroker/deployer-extended-database#db-download
  */
 task('db:download', function () {
     if (null === input()->getArgument('stage')) {
-        throw new \RuntimeException('The target instance is required for db:download command.', 1488143750580);
+        throw new GracefulShutdownException('The target instance is required for db:download command.', 1488143750580);
     }
     $dumpCode = (new ConsoleUtility())->optionRequired('dumpcode', input());
     $rsyncUtility = new RsyncUtility();