+++ deployer/db/task/db_download.php
@@ -25,4 +25,4 @@ task('db:download', function () {
         escapeshellarg($rsyncUtility->getHostWithDbStoragePath(get('target_stage'))),
         escapeshellarg($fileUtility->normalizeFolder(get('db_storage_path_current')))
     ), 0);
-})->desc('Download the database dumps with given dumpcode from target to current database dumps storage.');
+})->desc('Download the database dumps with given dumpcode from target to current database dumps storage');