+++ deployer/db/task/db_pull.php
@@ -23,4 +23,4 @@ task('db:pull', function () {
     runLocally($dl . ' db:import --dumpcode=' . $dumpCode . ' ' . $verbosity, 0);
     runLocally($dl . ' db:compress --dumpcode=' . $dumpCode . ' ' . $verbosity, 0);
     runLocally($dl . ' db:dumpclean' . $verbosity, 0);
-})->desc('Synchronize database from target instance to current instance.');
+})->desc('Synchronize database from target instance to current instance');