+++ deployer/db/task/db_truncate.php
@@ -41,4 +41,4 @@ task('db:truncate', function () {
         $activePath = get('deploy_path') . '/' . (test('[ -L {{deploy_path}}/release ]') ? 'release' : 'current');
         run('cd ' . $activePath . ' && {{bin/php}} {{bin/deployer}} db:truncate ' . $verbosity);
     }
-})->desc('Truncate tables defined in "truncate_tables" variable.');
+})->desc('Truncate tables defined in "truncate_tables" variable');