+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ master
 1) [FEATURE] Add option exportTaskAddIgnoreTablesToStructureDump to allow to add ignore tables when exporting structure.
 2) [FEATURE] Add option importTaskDoNotDropAllTablesBeforeImport to prevent dropping all tables before import.
 3) [TASK] Add vendor and composer.lock to .gitignore.
+4) [FEATURE][BREAKING] Implement sourcebroker/deployer-instance for instance management.
 
 7.0.2
 ~~~~~