+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ master
 2) [FEATURE] Add option importTaskDoNotDropAllTablesBeforeImport to prevent dropping all tables before import.
 3) [TASK] Add vendor and composer.lock to .gitignore.
 4) [FEATURE][BREAKING] Implement sourcebroker/deployer-instance for instance management.
+5) [BUGFIX] Remove colon from file names because if Windows compatibility.
 
 7.0.2
 ~~~~~