+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ master
 3) [TASK] Add vendor and composer.lock to .gitignore.
 4) [FEATURE][BREAKING] Implement sourcebroker/deployer-instance for instance management.
 5) [BUGFIX] Remove colon from file names because if Windows compatibility.
+6) [TASK] Replace RuntimeException with GracefulShutdownException.
 
 7.0.2
 ~~~~~